@@ -1,15 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:todolist/bloc/blocs/user_bloc_provider.dart';
-import 'package:todolist/bloc/resources/repository.dart';
 import 'package:todolist/models/subtasks.dart';
 import 'package:todolist/widgets/task_widgets/subtask_container_widget.dart';
 
 class SubtaskListTab extends StatefulWidget {
   final SubTaskBloc subTaskBloc;
-  final Repository repository;
   final taskKey;
 
-  SubtaskListTab(this.repository, this.taskKey, this.subTaskBloc);
+  SubtaskListTab(this.taskKey, this.subTaskBloc);
 
   @override
   _SubtaskListTabState createState() => _SubtaskListTabState();
@@ -38,7 +36,7 @@ class _SubtaskListTabState extends State<SubtaskListTab> {
       ),
       body: Container(
         child: SubtaskContainerWidget(
-            taskKey: widget.taskKey, repository: widget.repository),
+            taskKey: widget.taskKey),
       ),
     );
   }
