@@ -2,6 +2,7 @@
 
 import 'package:flutter/cupertino.dart';
 import 'package:todolist/models/group.dart';
+import 'package:todolist/models/groupmember.dart';
 import 'package:todolist/models/subtasks.dart';
 
 import '../resources/repository.dart';
@@ -10,9 +11,16 @@ import 'package:todolist/models/user.dart';
 import 'package:todolist/models/tasks.dart';
 
 class UserBloc {
-  final _repository = Repository();
   final _userGetter = PublishSubject<User>();
   User _user;
+
+  UserBloc._privateConstructor();
+
+  static final UserBloc _instance = UserBloc._privateConstructor();
+
+  factory UserBloc() {
+    return _instance;
+  }
 
   Observable<User> get getUser => _userGetter.stream;
 
@@ -22,18 +30,18 @@ class UserBloc {
 
   registerUser(String username, String password, String email, String firstname,
       String lastname, String phonenumber, ImageProvider avatar) async {
-    _user = await _repository.registerUser(
+    _user = await repository.registerUser(
         username, password, email, firstname, lastname, phonenumber, avatar);
     _userGetter.sink.add(_user);
   }
 
   signinUser(String username, String password, String apiKey) async {
     try {
-      _user = await _repository.signinUser(username, password, apiKey);
+      _user = await repository.signinUser(username, password, apiKey);
     } catch (e) {
       throw Exception(e.message);
-    } 
-      _userGetter.sink.add(_user);
+    }
+    _userGetter.sink.add(_user);
   }
 
   updateUserProfile(
@@ -47,7 +55,7 @@ class UserBloc {
       ImageProvider avatar,
       String apiKey) async {
     try {
-      _user = await _repository.updateUserProfile(currentPassword, newPassword,
+      _user = await repository.updateUserProfile(currentPassword, newPassword,
           email, username, firstname, lastname, phonenumber, avatar, apiKey);
     } catch (e) {
       throw Exception(e.message);
@@ -59,34 +67,66 @@ class UserBloc {
   }
 }
 
-class GroupBloc{
-  final _repository = Repository();
+class GroupBloc {
   final _groupSubject = BehaviorSubject<List<Group>>();
   String _apiKey;
-
-  var _groups = <Group>[];
-
-  GroupBloc(String apiKey){
-    this._apiKey = apiKey;
-    _updateGroups().then((_){
+  List<Group> _groups = <Group>[];
+  
+  void setApiKey() async {
+    _apiKey = await repository.getApiKey();
+    _updateGroups().then((_) {
       _groupSubject.add(_groups);
     });
   }
+
+  GroupBloc._privateConstructor() {
+    repository.getApiKey().then((apikey) {
+      if (apikey.isNotEmpty) {
+        this._apiKey = apikey;
+        _updateGroups().then((_) {
+          _groupSubject.add(_groups);
+        });
+      }
+    });
+  }
+
+  static final GroupBloc _instance = GroupBloc._privateConstructor();
+
+  factory GroupBloc() {
+    return _instance;
+  }
   /* Testing GroupBloc 
   Future<List<Group>> testGroupList() async{
-    return await _repository.getUserGroups(_apiKey);
+    return await repository.getUserGroups(_apiKey);
   } 
   */
 
   Stream<List<Group>> get getGroups => _groupSubject.stream;
-  Future<Null> _updateGroups() async{
-    _groups = await _repository.getUserGroups(_apiKey);
+  Future<Null> _updateGroups() async {
+    _groups = await repository.getUserGroups(_apiKey);
+  }
+}
+
+class GroupMemberBloc {
+  final _groupMemberSubject = BehaviorSubject<List<GroupMember>>();
+  String groupKey;
+  var _groupMembers = <GroupMember>[];
+
+
+  GroupMemberBloc(String groupKey) {
+    this.groupKey = groupKey;
+    _updateGroupMembers(groupKey).then((_) {
+      _groupMemberSubject.add(_groupMembers);
+    });
   }
 
+  Stream<List<GroupMember>> get getGroupMembers => _groupMemberSubject.stream;
+  Future<Null> _updateGroupMembers(String groupKey) async {
+    _groupMembers = await repository.getGroupMembers(groupKey);
+  }
 }
 
 class TaskBloc {
-  final _repository = Repository();
   final _taskSubject = BehaviorSubject<List<Task>>();
   String _apiKey;
 
@@ -101,12 +141,11 @@ class TaskBloc {
 
   Stream<List<Task>> get getTasks => _taskSubject.stream;
   Future<Null> _updateTasks() async {
-    _tasks = await _repository.getUserTasks(_apiKey);
+    _tasks = await repository.getUserTasks(_apiKey);
   }
 }
 
 class SubTaskBloc {
-  final _repository = Repository();
   final _subTaskSubject = BehaviorSubject<List<SubTask>>();
   String taskKey;
 
@@ -121,8 +160,10 @@ class SubTaskBloc {
 
   Stream<List<SubTask>> get getSubTasks => _subTaskSubject.stream;
   Future<Null> _updateSubTasks(String taskKey) async {
-    _subTasks = await _repository.getSubTasks(taskKey);
+    _subTasks = await repository.getSubTasks(taskKey);
   }
 }
 
 final userBloc = UserBloc();
+final groupBloc = GroupBloc();
+//final taskBloc = TaskBloc();
